requirements.txt
Flask
Flask-MySQLdb
Flask-Cors
Werkzeug
mysql-connector-python
